Enforce float tensors, add HPT config, update ensemble args
Key changes to fix dtype issues and wiring for ensemble runs and hyperparameter tuning:

- Cast encoder/decoder/target tensors to float() in EnsembleHelpers and EnsembleOutput to avoid dtype mismatches during training/evaluation.
- Update EnsembleMain.main signature to accept epochs and n_models and adjust _DataLoader call accordingly.
- Pass logger into RunEnsemble/EnsembleModel call from Main and remove redundant info logs in the full run path.
- Change torch.load(..., weights_only=False) in HPTHelpers to load full dataset for tuning.
- Add NewModelFolder/Files/HPTTuning.json containing tuned hyperparameters.
- Add two ensemble checkpoint files and update model checkpoint, evaluation plots and README to reflect new results.
- Clean up .gitignore entries (remove duplicates) and align argument docs formatting in run.sh.

These changes address runtime dtype errors, enable configurable ensemble training, and include hyperparameter tuning outputs and artifacts.

+++ NewModelFolder/Ensemble/EnsembleHelpers.py
@@ -155,7 +155,7 @@ def _TrainEnsemble(n_models, epochs, train_loader, val_loader, save_dir, config,
             epoch_loss = 0
 
             for enc, dec, tgt in train_loader:
-                enc, dec, tgt = enc.to(device), dec.to(device), tgt.to(device)
+                enc, dec, tgt = enc.to(device).float(), dec.to(device).float(), tgt.to(device).float()
 
                 optimizer.zero_grad()
                 mu, log_var = model(enc, dec)
@@ -172,7 +172,7 @@ def _TrainEnsemble(n_models, epochs, train_loader, val_loader, save_dir, config,
             val_loss_epoch = 0
             with torch.no_grad():
                 for enc, dec, tgt in val_loader:
-                    enc, dec, tgt = enc.to(device), dec.to(device), tgt.to(device)
+                    enc, dec, tgt = enc.to(device).float(), dec.to(device).float(), tgt.to(device).float()
                     mu, log_var = model(enc, dec)
                     loss = criterion(mu, tgt, torch.exp(log_var))
                     val_loss_epoch += loss.item() * enc.size(0)
